Set verify option from request to True
This is due the fact that this option has no effect on the actual mtls
certificate presented. We present our mtls_cert and it will be accepted
by the server or not (it will check the chain).

The verify is here for the main SSL certificate from the server, which
in our case is normally a letsencrypt certificate. We need to validate
that and not use our own ca for that unless we use self-signed
certificates on this.